+++ server/app/api/v1/__init__.py
@@ -4,8 +4,6 @@
 from fastapi import APIRouter
 
 from server.app.api.v1.artifacts import router as artifacts_router
-from server.app.api.v1.executions import router as executions_router
-from server.app.api.v1.lineage import router as lineage_router
 from server.app.api.v1.metadata import router as metadata_router
 from server.app.api.v1.pipelines import router as pipelines_router
 from server.app.api.v1.servers import router as servers_router
@@ -18,9 +16,7 @@ api_router = APIRouter()
 api_router.include_router(pipelines_router)
 api_router.include_router(metadata_router)
 api_router.include_router(servers_router)
-api_router.include_router(executions_router)
 api_router.include_router(artifacts_router)
-api_router.include_router(lineage_router)
 api_router.include_router(schedules_router)
 api_router.include_router(ui_actions_router)
 api_router.include_router(env_router)
